requirements.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.3.0
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
autopep8==1.5.2
beautifulsoup4==4.8.2
bson==0.5.8
cachetools==4.1.1
certifi==2020.4.5.1
cffi==1.14.0
channels==2.1.2
channels-api==0.4.1
channels-redis==2.4.2
chardet==3.0.4
cmsplugin-pagedown==0.0.2
colorama==0.4.3
configparser==5.0.0
constantly==15.1.0
cryptography==3.2.1
daphne==2.5.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.39.1
django-anonymous-notification==1.2.1
django-appconf==1.0.4
django-braces==1.14.0
django-classy-tags==2.0.0
django-crispy-forms==1.9.0
django-googledrive-storage==1.5.1
django-libs==2.0.3
django-mailer==2.0.1
django-markdown-deux==1.0.5
django-model-utils==4.0.0
django-notifications==0.1.dev0
django-notifications-hq==1.6.0
django-online-users==0.3
django-pagedown==2.0.3
django-rest-framework==0.1.0
django-sekizai==1.1.0
django-star-ratings==0.9.1
django-user-accounts==3.0.2
django-webpush==0.3.2
djangorestframework==3.11.0
EasyProcess==0.2.10
facebook-ads-api==0.3.0
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
gunicorn==20.0.4
h2==3.2.0
hiredis==1.1.0
hpack==3.0.0
http-ece==1.1.0
httpie==2.1.0
httpie-http2==0.0.1
httplib2==0.18.1
hyper==0.7.0
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.6.0
incremental==17.5.0
isort==5.6.4
jsonfield==3.1.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
markdown2==2.3.1
mccabe==0.6.1
msgpack==0.6.2
msgpack-python==0.5.6
mysql-connector==2.2.9
newsapi-python==0.2.6
notifications==0.3.2
oauth2client==4.1.3
oauthlib==3.1.0
Pillow==8.0.1
protobuf==3.13.0
py-vapid==1.7.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.2
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.6.0
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
pywebpush==1.9.4
redis==2.10.6
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.6
service-identity==18.1.0
six==1.15.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
swapper==1.1.2.post1
toml==0.10.2
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.25.11
wrapt==1.12.1
zipp==3.4.0
zope.interface==5.2.0
